fix, feat: unresolved which asset in session plan and refactored ops
