+++ phymes-etl/src/candle_ops/ops_processor.rs
@@ -1,9 +1,9 @@
 use super::{
     ops_config::CandleOpsConfig,
-    ops_functions::{chunk_documents, relative_similarity_scores, sort_scores_and_indices, join_inner},
     ops_service::CandleOpsService,
     ops_which::WhichCandleOps,
 };
+use crate::ops_functions::{chunk_documents::chunk_documents, relative_similarity_scores::relative_similarity_scores, sort_scores_and_indices::sort_scores_and_indices, join_inner::join_inner},
 use phymes_ai::candle_assets::device::device;
 use phymes_core::{
     metrics::{ArrowTaskMetricsSet, BaselineMetrics, HashMap},
